+++ lib/app.dart
@@ -3,7 +3,6 @@ import 'package:provider/provider.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
-import 'package:firebase_analytics/observer.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import 'package:first_app/models/appstate.dart';
@@ -12,8 +11,8 @@ import 'package:first_app/ui/pages/login/index.dart';
 import 'package:first_app/ui/theme/style.dart';
 
 // Using async functions must be done from an async function
-Future<Widget> getApp({bool mock = false}) async {
-  var analytics = FirebaseAnalytics();
+Future<Widget> getApp({bool mock = false, bool web = false}) async {
+  var analytics = FirebaseAnalytics.instance;
   // Wrap a StatelessWidget (ProviderApp) in a ChangeNotifierProvider to trigger rebuild of the
   // entire MaterialApp when app state, like locale, changes
   return ChangeNotifierProvider.value(
@@ -21,7 +20,8 @@ Future<Widget> getApp({bool mock = false}) async {
           prefs: await SharedPreferences.getInstance(),
           analytics: analytics,
           messaging: FirebaseMessaging.instance,
-          mock: mock),
+          mock: mock,
+          web: web),
       child: ProviderApp(analytics: analytics));
 }