+++ lib/main.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'app.dart';
+import 'firebase_options.dart';
 
 void main() async {
   // A breaking change in the platform messaging, as of Flutter 1.12.13+hotfix.5,
@@ -12,24 +13,23 @@ void main() async {
   // https://groups.google.com/forum/#!msg/flutter-announce/sHAL2fBtJ1Y/mGjrKH3dEwAJ
   WidgetsFlutterBinding.ensureInitialized();
 
-  await Firebase.initializeApp();
+  await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
+
+  // If we run on web, do not use Crashlytics (not supported on web yet)
   if (kIsWeb) {
     FlutterError.onError = (FlutterErrorDetails details) {
       FlutterError.dumpErrorToConsole(details);
     };
-  } else {
-    // Pass all uncaught errors from the framework to Crashlytics.
-    FlutterError.onError = FirebaseCrashlytics.instance.recordFlutterError;
-  }
-
-  // If we run on web, do not use Crashlytics (not supported on web yet)
-  if (kIsWeb) {
-    runApp(await getApp(mock: true));
+    runApp(await getApp(web: true));
   } else {
     // Use dart zone to define Crashlytics as error handler for errors
     // that occur outside runApp
     runZonedGuarded<Future<void>>(() async {
+      // Pass all uncaught errors from the framework to Crashlytics.
+      FlutterError.onError = FirebaseCrashlytics.instance.recordFlutterError;
       runApp(await getApp());
-    }, FirebaseCrashlytics.instance.recordError);
+    },
+        (error, stack) =>
+            FirebaseCrashlytics.instance.recordError(error, stack));
   }
 }